requirements.txt
# CommonForms API - Requirements
# Generated for Python 3.12.7
# Install with: pip install -r requirements.txt

# FastAPI and ASGI Server
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.3
bcrypt==4.2.0

# Settings Management
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic-core==2.23.4

# PDF Processing
commonforms==0.1.2
PyPDF2==3.0.1
PyMuPDF==1.24.13
Pillow==10.4.0

# Machine Learning & GPU
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
# For CUDA 12.4 support (Windows/Linux):
# torch==2.5.1+cu124
# torchvision==0.20.1+cu124
# torchaudio==2.5.1+cu124
# Install with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124

# Utilities
python-dotenv==1.0.1
typing-extensions==4.12.2

# HTTP Client (for testing)
httpx==0.27.2
requests==2.32.3

# Additional Dependencies
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.0
email-validator==2.2.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.2
orjson==3.10.10
packaging==24.2
pluggy==1.5.0
pyasn1==0.6.1
pycparser==2.22
pydantic_core==2.23.4
pytest==8.3.3
python-dotenv==1.0.1
PyYAML==6.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
ujson==5.10.0
urllib3==2.2.3
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1

# Development Dependencies (optional)
# pytest==8.3.3
# pytest-asyncio==0.24.0
# black==24.10.0
# flake8==7.1.1
# mypy==1.13.0
